requirements.txt
PyGithub
GitPython==3.1.32
python_dateutil==2.8.2
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
tqdm==4.64.0